refactor: rebuild the main with the correct SmallCarCard component and minor issues
