+++ src/App.vue
@@ -16,7 +16,6 @@ export default {
     <router-link to="/danielView">Daniels View</router-link>|
     <router-link to="/mainView">Main Page</router-link>|
     <router-link to="/KirstensView">KirstensView</router-link>|
-    <router-link to="/CarsView">Verfügbare Autos</router-link> |
     <router-link to="/AddNewCarView">Neues Auto erfassen</router-link>|
     <router-link to="/testUserManagementView"
       >User/Data Managment Test Bereich</router-link
