--- src/components/SmallCarCard.vue
@@ -1,335 +1,0 @@
-<template>
-  <article class="car-card__wrapper">
-    <img
-      src="https://ais-cf.tvnow.de/tvnow/movie/5429467/1800x0/folge-vom-23042023.jpg"
-      alt="Das Auto des Users"
-    />
-
-    <section class="car-card__important-info">
-      <section class="car-card__basic-information">
-        <span class="car-card__adress">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="100%"
-            height="100%"
-            fill="currentColor"
-            class="bi bi-geo-alt-fill"
-            viewBox="0 0 16 16"
-          >
-            <path
-              d="M8 16s6-5.686 6-10A6 6 0 0 0 2 6c0 4.314 6 10 6 10zm0-7a3 3 0 1 1 0-6 3 3 0 0 1 0 6z"
-            />
-          </svg>
-          <p>01097</p>
-          <p>/</p>
-          <p>Dresden, Neustadt</p>
-        </span>
-        <span class="car-card__seat-icon-wrapper">
-          <svg
-            version="1.1"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            viewBox="0, 0, 14.173, 19.843"
-          >
-            <g id='"_-Layer_4'>
-              <g>
-                <path
-                  d="M12.687,13.188 L12.944,13.202 C13.614,13.38 13.687,16.612 12.87,17.787 L12.401,18.391 L10.517,18.391 C9.666,18.407 8.699,18.459 7.859,18.391 L6.58,18.391 L6.58,18.391 L6.314,18.391 C5.474,18.459 4.507,18.407 3.656,18.391 L1.772,18.391 L1.303,17.787 C0.486,16.612 0.56,13.38 1.229,13.202 L1.486,13.188 C2.327,13.441 2.68,14.262 2.961,15.024 C4.246,14.478 5.716,14.522 7.087,14.478 C8.457,14.522 9.928,14.478 11.213,15.024 C11.493,14.262 11.846,13.441 12.687,13.188 z"
-                />
-                <path
-                  d="M10.458,5.226 C10.969,5.226 11.382,5.644 11.382,6.16 L11.662,12.989 C10.916,13.822 11.16,13.386 10.834,14.236 C9.669,13.786 8.325,13.807 7.087,13.749 C5.834,13.774 4.542,13.832 3.35,14.264 C3.013,13.386 3.264,13.829 2.511,12.989 L2.791,6.16 C2.791,5.644 3.205,5.226 3.715,5.226 L10.458,5.226 z"
-                />
-                <path
-                  d="M4.141,4.451 L4.079,4.1 L4.079,2.562 C4.079,1.93 4.528,1.417 5.081,1.417 L9.093,1.417 C9.646,1.417 10.094,1.93 10.094,2.562 L10.094,4.1 L10.032,4.451 L4.141,4.451 z"
-                />
-                <path
-                  d="M10.594,5.726 C10.616,7.243 4.149,12.292 4.64,14.546"
-                  fill-opacity="0"
-                  stroke-width="1"
-                />
-              </g>
-            </g>
-          </svg>
-          <svg
-            version="1.1"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            viewBox="0, 0, 14.173, 19.843"
-          >
-            <g id='"_-Layer_4'>
-              <g>
-                <path
-                  d="M12.687,13.188 L12.944,13.202 C13.614,13.38 13.687,16.612 12.87,17.787 L12.401,18.391 L10.517,18.391 C9.666,18.407 8.699,18.459 7.859,18.391 L6.58,18.391 L6.58,18.391 L6.314,18.391 C5.474,18.459 4.507,18.407 3.656,18.391 L1.772,18.391 L1.303,17.787 C0.486,16.612 0.56,13.38 1.229,13.202 L1.486,13.188 C2.327,13.441 2.68,14.262 2.961,15.024 C4.246,14.478 5.716,14.522 7.087,14.478 C8.457,14.522 9.928,14.478 11.213,15.024 C11.493,14.262 11.846,13.441 12.687,13.188 z"
-                />
-                <path
-                  d="M10.458,5.226 C10.969,5.226 11.382,5.644 11.382,6.16 L11.662,12.989 C10.916,13.822 11.16,13.386 10.834,14.236 C9.669,13.786 8.325,13.807 7.087,13.749 C5.834,13.774 4.542,13.832 3.35,14.264 C3.013,13.386 3.264,13.829 2.511,12.989 L2.791,6.16 C2.791,5.644 3.205,5.226 3.715,5.226 L10.458,5.226 z"
-                />
-                <path
-                  d="M4.141,4.451 L4.079,4.1 L4.079,2.562 C4.079,1.93 4.528,1.417 5.081,1.417 L9.093,1.417 C9.646,1.417 10.094,1.93 10.094,2.562 L10.094,4.1 L10.032,4.451 L4.141,4.451 z"
-                />
-                <path
-                  d="M10.594,5.726 C10.616,7.243 4.149,12.292 4.64,14.546"
-                  fill-opacity="0"
-                  stroke-width="1"
-                />
-              </g>
-            </g>
-          </svg>
-          <svg
-            version="1.1"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            viewBox="0, 0, 14.173, 19.843"
-          >
-            <g id='"_-Layer_4'>
-              <g>
-                <path
-                  d="M12.687,13.188 L12.944,13.202 C13.614,13.38 13.687,16.612 12.87,17.787 L12.401,18.391 L10.517,18.391 C9.666,18.407 8.699,18.459 7.859,18.391 L6.58,18.391 L6.58,18.391 L6.314,18.391 C5.474,18.459 4.507,18.407 3.656,18.391 L1.772,18.391 L1.303,17.787 C0.486,16.612 0.56,13.38 1.229,13.202 L1.486,13.188 C2.327,13.441 2.68,14.262 2.961,15.024 C4.246,14.478 5.716,14.522 7.087,14.478 C8.457,14.522 9.928,14.478 11.213,15.024 C11.493,14.262 11.846,13.441 12.687,13.188 z"
-                />
-                <path
-                  d="M10.458,5.226 C10.969,5.226 11.382,5.644 11.382,6.16 L11.662,12.989 C10.916,13.822 11.16,13.386 10.834,14.236 C9.669,13.786 8.325,13.807 7.087,13.749 C5.834,13.774 4.542,13.832 3.35,14.264 C3.013,13.386 3.264,13.829 2.511,12.989 L2.791,6.16 C2.791,5.644 3.205,5.226 3.715,5.226 L10.458,5.226 z"
-                />
-                <path
-                  d="M4.141,4.451 L4.079,4.1 L4.079,2.562 C4.079,1.93 4.528,1.417 5.081,1.417 L9.093,1.417 C9.646,1.417 10.094,1.93 10.094,2.562 L10.094,4.1 L10.032,4.451 L4.141,4.451 z"
-                />
-                <path
-                  d="M10.594,5.726 C10.616,7.243 4.149,12.292 4.64,14.546"
-                  fill-opacity="0"
-                  stroke-width="1"
-                />
-              </g>
-            </g>
-          </svg>
-          <svg
-            version="1.1"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            viewBox="0, 0, 14.173, 19.843"
-          >
-            <g id='"_-Layer_4'>
-              <g>
-                <path
-                  d="M12.687,13.188 L12.944,13.202 C13.614,13.38 13.687,16.612 12.87,17.787 L12.401,18.391 L10.517,18.391 C9.666,18.407 8.699,18.459 7.859,18.391 L6.58,18.391 L6.58,18.391 L6.314,18.391 C5.474,18.459 4.507,18.407 3.656,18.391 L1.772,18.391 L1.303,17.787 C0.486,16.612 0.56,13.38 1.229,13.202 L1.486,13.188 C2.327,13.441 2.68,14.262 2.961,15.024 C4.246,14.478 5.716,14.522 7.087,14.478 C8.457,14.522 9.928,14.478 11.213,15.024 C11.493,14.262 11.846,13.441 12.687,13.188 z"
-                />
-                <path
-                  d="M10.458,5.226 C10.969,5.226 11.382,5.644 11.382,6.16 L11.662,12.989 C10.916,13.822 11.16,13.386 10.834,14.236 C9.669,13.786 8.325,13.807 7.087,13.749 C5.834,13.774 4.542,13.832 3.35,14.264 C3.013,13.386 3.264,13.829 2.511,12.989 L2.791,6.16 C2.791,5.644 3.205,5.226 3.715,5.226 L10.458,5.226 z"
-                />
-                <path
-                  d="M4.141,4.451 L4.079,4.1 L4.079,2.562 C4.079,1.93 4.528,1.417 5.081,1.417 L9.093,1.417 C9.646,1.417 10.094,1.93 10.094,2.562 L10.094,4.1 L10.032,4.451 L4.141,4.451 z"
-                />
-                <path
-                  d="M10.594,5.726 C10.616,7.243 4.149,12.292 4.64,14.546"
-                  fill-opacity="0"
-                  stroke-width="1"
-                />
-              </g>
-            </g>
-          </svg>
-          <svg
-            version="1.1"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            viewBox="0, 0, 14.173, 19.843"
-          >
-            <g id='"_-Layer_4'>
-              <g>
-                <path
-                  d="M12.687,13.188 L12.944,13.202 C13.614,13.38 13.687,16.612 12.87,17.787 L12.401,18.391 L10.517,18.391 C9.666,18.407 8.699,18.459 7.859,18.391 L6.58,18.391 L6.58,18.391 L6.314,18.391 C5.474,18.459 4.507,18.407 3.656,18.391 L1.772,18.391 L1.303,17.787 C0.486,16.612 0.56,13.38 1.229,13.202 L1.486,13.188 C2.327,13.441 2.68,14.262 2.961,15.024 C4.246,14.478 5.716,14.522 7.087,14.478 C8.457,14.522 9.928,14.478 11.213,15.024 C11.493,14.262 11.846,13.441 12.687,13.188 z"
-                />
-                <path
-                  d="M10.458,5.226 C10.969,5.226 11.382,5.644 11.382,6.16 L11.662,12.989 C10.916,13.822 11.16,13.386 10.834,14.236 C9.669,13.786 8.325,13.807 7.087,13.749 C5.834,13.774 4.542,13.832 3.35,14.264 C3.013,13.386 3.264,13.829 2.511,12.989 L2.791,6.16 C2.791,5.644 3.205,5.226 3.715,5.226 L10.458,5.226 z"
-                />
-                <path
-                  d="M4.141,4.451 L4.079,4.1 L4.079,2.562 C4.079,1.93 4.528,1.417 5.081,1.417 L9.093,1.417 C9.646,1.417 10.094,1.93 10.094,2.562 L10.094,4.1 L10.032,4.451 L4.141,4.451 z"
-                />
-                <path
-                  d="M10.594,5.726 C10.616,7.243 4.149,12.292 4.64,14.546"
-                  fill-opacity="0"
-                  stroke-width="1"
-                />
-              </g>
-            </g>
-          </svg>
-          <svg
-            version="1.1"
-            xmlns="http://www.w3.org/2000/svg"
-            xmlns:xlink="http://www.w3.org/1999/xlink"
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            viewBox="0, 0, 14.173, 19.843"
-          >
-            <g id='"_-Layer_4'>
-              <g>
-                <path
-                  d="M12.687,13.188 L12.944,13.202 C13.614,13.38 13.687,16.612 12.87,17.787 L12.401,18.391 L10.517,18.391 C9.666,18.407 8.699,18.459 7.859,18.391 L6.58,18.391 L6.58,18.391 L6.314,18.391 C5.474,18.459 4.507,18.407 3.656,18.391 L1.772,18.391 L1.303,17.787 C0.486,16.612 0.56,13.38 1.229,13.202 L1.486,13.188 C2.327,13.441 2.68,14.262 2.961,15.024 C4.246,14.478 5.716,14.522 7.087,14.478 C8.457,14.522 9.928,14.478 11.213,15.024 C11.493,14.262 11.846,13.441 12.687,13.188 z"
-                />
-                <path
-                  d="M10.458,5.226 C10.969,5.226 11.382,5.644 11.382,6.16 L11.662,12.989 C10.916,13.822 11.16,13.386 10.834,14.236 C9.669,13.786 8.325,13.807 7.087,13.749 C5.834,13.774 4.542,13.832 3.35,14.264 C3.013,13.386 3.264,13.829 2.511,12.989 L2.791,6.16 C2.791,5.644 3.205,5.226 3.715,5.226 L10.458,5.226 z"
-                />
-                <path
-                  d="M4.141,4.451 L4.079,4.1 L4.079,2.562 C4.079,1.93 4.528,1.417 5.081,1.417 L9.093,1.417 C9.646,1.417 10.094,1.93 10.094,2.562 L10.094,4.1 L10.032,4.451 L4.141,4.451 z"
-                />
-                <path
-                  d="M10.594,5.726 C10.616,7.243 4.149,12.292 4.64,14.546"
-                  fill-opacity="0"
-                  stroke-width="1"
-                />
-              </g>
-            </g>
-          </svg>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="100%"
-            height="100%"
-            fill="currentColor"
-            class="bi bi-plus plus"
-            viewBox="0 0 16 16"
-          >
-            <path
-              d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z"
-            />
-          </svg>
-        </span>
-      </section>
-      <section class="car-card__icon-wrapper" style="--svg-width: 2.4rem">
-        <IconType :carType="'Kleinwagen'" class="icon"></IconType>
-        <FuelType :fuelType="'Gas'" class="icon"></FuelType>
-        <TrunkType :trunkType="'S'" class="icon"></TrunkType>
-        <DifferentType :differentType="'Isofix'" class="icon"></DifferentType>
-        <DifferentType :differentType="'Nicht-Raucher'" class="icon">
-        </DifferentType>
-        <DifferentType :differentType="'Tiere'" class="icon"> </DifferentType>
-      </section>
-    </section>
-  </article>
-</template>
-
-<script>
-import IconType from "@/components/icon-type/IconType.vue";
-import FuelType from "@/components/icon-type/fuelType.vue";
-import TrunkType from "@/components/icon-type/TrunkType.vue";
-import DifferentType from "@/components/icon-type/DifferentTypes.vue";
-
-export default {
-  components: {
-    IconType,
-    FuelType,
-    TrunkType,
-    DifferentType,
-  },
-};
-</script>
-
-<style scoped>
-.car-card__wrapper {
-  width: 100%;
-  height: 18rem;
-  position: relative;
-  margin-bottom: 1rem;
-  border-radius: 1rem;
-  overflow: hidden;
-  border-top: 0.125rem solid var(--clr-img-top);
-  box-shadow: 10px 10px 10px var(--card-shadow);
-}
-
-img {
-  border-radius: 1rem 1rem 0 0;
-  width: 100%;
-  filter: grayscale(var(--filter-img));
-}
-
-.car-card__basic-information {
-  width: 95%;
-  margin-inline: auto;
-
-  display: flex;
-}
-.car-card__important-info {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  width: 100%;
-  padding-block: 1.2rem;
-  border-top: 1px solid var(--clr-brd-top);
-  background: var(--surface-light);
-  border-radius: 1rem;
-
-  box-shadow: 0 -10px 15px var(--nav-shadow);
-}
-.car-card__adress {
-  width: 95%;
-  margin-inline: auto;
-  height: 2rem;
-  display: flex;
-  align-items: center;
-}
-.car-card__adress p {
-  font-weight: 300;
-  font-size: 0.8rem;
-  color: var(--clr-plz-car, rgb(179, 179, 179));
-}
-.car-card__adress > svg {
-  width: 1.2rem;
-  aspect-ratio: 1;
-  fill: var(--clr-plz-car, rgb(179, 179, 179));
-  margin-right: 0.5rem;
-}
-.icon {
-  width: 3rem;
-  aspect-ratio: 1;
-  margin-inline: 0.4rem;
-}
-.car-card__icon-wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 0.2rem;
-  padding-left: -0.3rem;
-  padding-block: 0.9rem 0rem;
-
-  width: 90%;
-  margin-inline: auto;
-  margin-top: 0.5rem;
-  border-top: 1px solid var(--clr-brd);
-}
-
-.car-card__seat-icon-wrapper {
-  align-items: center;
-  display: flex;
-}
-.car-card__seat-icon-wrapper > svg {
-  fill: var(--clr-seat-def);
-  width: 1rem;
-}
-.car-card__seat-icon-wrapper > .plus {
-  fill: var(--clr-state-icons);
-  width: 2rem;
-  aspect-ratio: 1;
-}
-</style>
